NAP-009: Added volume and registry related functions
