+++ docs/html/index.html
@@ -69,6 +69,17 @@ el.replaceWith(d);
 <dd>
 <div class="desc"><p>Manage OpenSearch releated function on MS …</p></div>
 </dd>
+<dt><code class="name"><a title="nerve_lib.manage_registry" href="manage_registry.html">nerve_lib.manage_registry</a></code></dt>
+<dd>
+<div class="desc"><p>Manage Docker registry images on nodes from MS.
+This module provides the InternalRegistry class for managing Docker
+registry images and their tags.
+It …</p></div>
+</dd>
+<dt><code class="name"><a title="nerve_lib.manage_volumes" href="manage_volumes.html">nerve_lib.manage_volumes</a></code></dt>
+<dd>
+<div class="desc"><p>Manage Docker volumes and unused images on nodes from MS …</p></div>
+</dd>
 <dt><code class="name"><a title="nerve_lib.manage_workloads" href="manage_workloads.html">nerve_lib.manage_workloads</a></code></dt>
 <dd>
 <div class="desc"><p>Manage workloads on MS or LocalUi …</p></div>
@@ -96,6 +107,8 @@ el.replaceWith(d);
 <li><code><a title="nerve_lib.manage_node" href="manage_node.html">nerve_lib.manage_node</a></code></li>
 <li><code><a title="nerve_lib.manage_notifications" href="manage_notifications.html">nerve_lib.manage_notifications</a></code></li>
 <li><code><a title="nerve_lib.manage_open_search" href="manage_open_search.html">nerve_lib.manage_open_search</a></code></li>
+<li><code><a title="nerve_lib.manage_registry" href="manage_registry.html">nerve_lib.manage_registry</a></code></li>
+<li><code><a title="nerve_lib.manage_volumes" href="manage_volumes.html">nerve_lib.manage_volumes</a></code></li>
 <li><code><a title="nerve_lib.manage_workloads" href="manage_workloads.html">nerve_lib.manage_workloads</a></code></li>
 </ul>
 </li>
