+++ docs/markdown/manage_workloads.md
@@ -16,149 +16,8 @@ Example:
     >>>                     restart_policy="always")
     >>>     wl.provision_workload(wl_config, file_paths=["images/docker.tar"])
 
-Functions
----------
-
-`remove_unused_images_ms(self, dut_serial: type)`
-:   Remove unused images from node.
-    
-    Parameters
-    ----------
-    dut : type
-        reference to node handles (general_utils.NodeHandle).
-
 Classes
 -------
-
-`DockerVolumes(ms_handle: type)`
-:   Handle to Docker volumes on node from MS.
-    
-    Parameters
-    ----------
-    ms_handle : type
-        handle to manage_workloads.Workloads object.
-
-    ### Methods
-
-    `check_export_status(self, dut_serial, volume_name, retry_timeout=60)`
-    :   Check the status of the export operation.
-        
-        Parameters
-        ----------
-        dut_serial : type
-            reference to node serial number.
-        volume_name : str
-            Name of the volume.
-
-    `delete_all_volumes(self, dut_serial: str)`
-    :   Delete all docker volumes on a node.
-
-    `delete_volume(self, dut_serial: type, volume_name: str)`
-    :   Delete docker volume.
-        
-        Parameters
-        ----------
-        dut_serial: type
-            reference to node serial number.
-        volume_name: str
-            docker volume name.
-
-    `export_volume_data_ms(self, dut_serial, volume_name, export_timeout=30)`
-    :   Export data from a volume.
-        
-        Parameters
-        ----------
-        dut_serial : type
-            reference to node serial number.
-        volume_name : str
-            Name of the volume.
-        export_timeout : int, optional
-            Timeout for the export operation. The default is 30 seconds.
-
-    `get_volumes(self, dut_serial: type)`
-    :   Remove unused images from node.
-        
-        Parameters
-        ----------
-        dut_serial : type
-            reference to node serial number.
-
-    `import_volume_data_ms(self, dut_serial, volume_name, file, import_timeout=30)`
-    :   Import data to a volume.
-        
-        Parameters
-        ----------
-        volume_name : str
-            Name of the volume.
-        file : str
-            Path to the file to be imported.
-
-`InternalRegistry(ms_handle: type)`
-:   Handle to internal registry of a MS.
-    
-    Parameters
-    ----------
-    owner : type
-        handle to manage_workloads.Workloads object.
-    workload_name : str
-        selected workload name.
-    version : str, optional
-        selected workload version. If empty, the last version will be selected. The default is "".
-    release_version : str, optional
-        selected release version. If empty, the last version will be selected. The default is "".
-
-    ### Methods
-
-    `check_image_status(self, images: list[str] = [])`
-    :   Check image status.
-        
-        Parameters
-        ----------
-        images : list[str]
-            image name(s).
-        
-        Returns
-        -------
-        dict
-            MS API response containing image status.
-
-    `delete_image_tag(self, image: str, tag: str)`
-    :   Delete image tag.
-        
-        Parameters
-        ----------
-        image : str
-            image name.
-        tag : str
-            tag name.
-
-    `get_image_tags(self, image: str)`
-    :   Get image tags.
-        
-        Parameters
-        ----------
-        image : str
-            image name.
-        
-        Returns
-        -------
-        dict
-            MS API response containing image tags.
-
-    `get_registry_images(self, last: str = '', limit: int = 0)`
-    :   Get registry images.
-        
-        Parameters
-        ----------
-        last : str, optional
-            The starting point for fetching registry images. Default is "".
-        limit : int, optional
-            The maximum number of images to retrieve. Default is 0.
-        
-        Returns
-        -------
-        dict
-            MS API response containing registry images.
 
 `LocalWorkloads(node_handle: type)`
 :   Manage workloads using localUI of a node.
