+++ nerve_lib/manage_workloads.py
@@ -1298,7 +1298,7 @@ class _WorkloadVersion:  # noqa: PLR0904
         workload["versions"] = []
         if include_version:
             version = self._get_versions(selected_version=True)
-            workload["versions"] = [version]
+            workload["versions"] = version
 
         return workload
 
@@ -2101,274 +2101,3 @@ class _WorkloadVersion:  # noqa: PLR0904
                 accepted_status=[requests.codes.ok],
             )
             self._log.info("Version export started!")
-
-
-class InternalRegistry:
-    """Handle to internal registry of a MS.
-
-    Parameters
-    ----------
-    owner : type
-        handle to manage_workloads.Workloads object.
-    workload_name : str
-        selected workload name.
-    version : str, optional
-        selected workload version. If empty, the last version will be selected. The default is "".
-    release_version : str, optional
-        selected release version. If empty, the last version will be selected. The default is "".
-    """
-
-    def __init__(self, ms_handle: type):
-        self.ms = ms_handle
-
-    def get_registry_images(self, last: str = "", limit: int = 0):
-        """Get registry images.
-
-        Parameters
-        ----------
-        last : str, optional
-            The starting point for fetching registry images. Default is "".
-        limit : int, optional
-            The maximum number of images to retrieve. Default is 0.
-
-        Returns
-        -------
-        dict
-            MS API response containing registry images.
-        """
-        payload = {}
-        if last:
-            payload["last"] = last
-        if limit > 0:
-            payload["limit"] = limit
-
-        return self.ms.post(
-            "nerve/registry/images",
-            accepted_status=[requests.codes.ok],
-            json=payload,
-        ).json()
-
-    def get_image_tags(self, image: str):
-        """Get image tags.
-
-        Parameters
-        ----------
-        image : str
-            image name.
-
-        Returns
-        -------
-        dict
-            MS API response containing image tags.
-        """
-        payload = {"image": image}
-        return self.ms.post(
-            "nerve/registry/image-tags",
-            accepted_status=[requests.codes.ok],
-            json=payload,
-        ).json()
-
-    def check_image_status(self, images: list[str] = []):
-        """Check image status.
-
-        Parameters
-        ----------
-        images : list[str]
-            image name(s).
-
-        Returns
-        -------
-        dict
-            MS API response containing image status.
-        """
-        payload = {"images": images}
-        return self.ms.post(
-            "nerve/registry/check-images-status",
-            accepted_status=[requests.codes.ok],
-            json=payload,
-        ).json()
-
-    def delete_image_tag(self, image: str, tag: str):
-        """Delete image tag.
-
-        Parameters
-        ----------
-        image : str
-            image name.
-        tag : str
-            tag name.
-        """
-        payload = {"image": image, "tag": tag}
-        return self.ms.post(
-            "nerve/registry/delete-image-tag",
-            accepted_status=[requests.codes.ok],
-            json=payload,
-        ).json()
-
-
-def remove_unused_images_ms(self, dut_serial: type):
-    """Remove unused images from node.
-
-    Parameters
-    ----------
-    dut : type
-        reference to node handles (general_utils.NodeHandle).
-    """
-    return self.ms.delete(
-        f"nerve/v2/node/{dut_serial}/docker-resources/images/unused",
-        accepted_status=[requests.codes.no_content],
-    )
-
-
-class DockerVolumes:
-    """Handle to Docker volumes on node from MS.
-
-    Parameters
-    ----------
-    ms_handle : type
-        handle to manage_workloads.Workloads object.
-    """
-
-    def __init__(self, ms_handle: type):
-        self.ms = ms_handle
-
-    def get_volumes(self, dut_serial: type):
-        """Remove unused images from node.
-
-        Parameters
-        ----------
-        dut_serial : type
-            reference to node serial number.
-        """
-        return self.ms.get(
-            f"nerve/v2/node/{dut_serial}/docker-resources/volumes",
-            accepted_status=[requests.codes.ok],
-        )
-
-    def delete_volume(self, dut_serial: type, volume_name: str):
-        """Delete docker volume.
-
-        Parameters
-        ----------
-        dut_serial: type
-            reference to node serial number.
-        volume_name: str
-            docker volume name.
-        """
-        return self.ms.delete(
-            f"nerve/v2/node/{dut_serial}/docker-resources/volumes/{volume_name}",
-            accepted_status=[requests.codes.no_content],
-        )
-
-    # Delete all volumes on MS
-    def delete_all_volumes(self, dut_serial: str):
-        """Delete all docker volumes on a node."""
-        # Get a list of all volumes
-        response = self.get_volumes(dut_serial)
-
-        if response.status_code != requests.codes.ok:
-            raise RuntimeError(f"Failed to fetch volumes: {response.status_code} - {response.text}")
-
-        data = response.json()
-
-        # Check if the response is in the expected format
-        if "volumes" not in data or not isinstance(data["volumes"], list):
-            raise RuntimeError(f"Unexpected response format: {data}")
-
-        # List of volume names
-        volumes = [volume["name"] for volume in data["volumes"]]
-
-        # Delete all volumes from the list
-        for volume in volumes:
-            print(f"Deleting volume: {volume}")
-            self.delete_volume(dut_serial, volume)
-
-    # Import volume data
-    def import_volume_data_ms(self, dut_serial, volume_name, file, import_timeout=30):
-        """Import data to a volume.
-
-        Parameters
-        ----------
-        volume_name : str
-            Name of the volume.
-        file : str
-            Path to the file to be imported.
-        """
-        m_enc = MultipartEncoder({"file": (os.path.basename(file), open(file, "rb"), "form-data")})
-
-        self.ms.login()
-        sessionid = self.ms._add_header["sessionid"]
-
-        headers = {
-            "Connection": "close",
-            "Content-Type": m_enc.content_type,
-            "sessionId": sessionid,
-        }
-
-        return self.ms.post(
-            url=f"/nerve/v2/node/{dut_serial}/docker-resources/volumes/{volume_name}/import",
-            headers=headers,  # Add headers here
-            data=m_enc,
-            accepted_status=[requests.codes.ok],
-            timeout=(30, import_timeout),
-        )
-
-    # Export volume data
-    def export_volume_data_ms(self, dut_serial, volume_name, export_timeout=30):
-        """Export data from a volume.
-
-        Parameters
-        ----------
-        dut_serial : type
-            reference to node serial number.
-        volume_name : str
-            Name of the volume.
-        export_timeout : int, optional
-            Timeout for the export operation. The default is 30 seconds.
-        """
-        self.ms.login()
-        sessionid = self.ms._add_header["sessionid"]
-
-        headers = {
-            "Connection": "keep-alive",
-            "sessionId": sessionid,
-        }
-
-        return self.ms.post(
-            url=f"/nerve/v2/node/{dut_serial}/docker-resources/volumes/{volume_name}/export",
-            headers=headers,  # Add headers here
-            accepted_status=[requests.codes.no_content],
-            timeout=(30, export_timeout),
-        )
-
-    # Check export status
-    def check_export_status(self, dut_serial, volume_name, retry_timeout=60):
-        """Check the status of the export operation.
-
-        Parameters
-        ----------
-        dut_serial : type
-            reference to node serial number.
-        volume_name : str
-            Name of the volume.
-        """
-        start_time = time.time()
-
-        while time.time() - start_time < retry_timeout:
-            response = self.get_volumes(dut_serial).json()
-
-            for volume in response["volumes"]:
-                if volume["name"] != "compose_ms_registry_test":
-                    continue
-                for info in volume["backupInfo"]:
-                    if info["action"] != "export":
-                        continue
-                    if info["status"] == "COMPLETED":
-                        return info["backupName"]
-
-            time.sleep(10)  # wait 10 seconds before retrying
-
-        # If timeout is reached without completion
-        raise TimeoutError(
-            f"Export did not complete within {retry_timeout} seconds for volume '{volume_name}'."
-        )
